move delete manager tests
